test: isolate project parser coverage
